+++ probe-rs/src/probe/cmsisdap/commands/general/connect.rs
@@ -1,5 +1,4 @@
-use super::super::{Category, CmsisDapError, Request, Response, Result};
-use anyhow::{anyhow, Context};
+use super::super::{Category, Request, Response, SendError};
 
 #[derive(Clone, Copy, Debug)]
 pub enum ConnectRequest {
@@ -11,7 +10,7 @@ pub enum ConnectRequest {
 impl Request for ConnectRequest {
     const CATEGORY: Category = Category(0x02);
 
-    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize> {
+    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize, SendError> {
         buffer[offset] = *self as u8;
         Ok(1)
     }
@@ -25,17 +24,12 @@ pub enum ConnectResponse {
 }
 
 impl Response for ConnectResponse {
-    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self> {
+    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self, SendError> {
         match buffer[offset] {
             0 => Ok(ConnectResponse::InitFailed),
             1 => Ok(ConnectResponse::SuccessfulInitForSWD),
             2 => Ok(ConnectResponse::SuccessfulInitForJTAG),
-            _ => Err(anyhow!(CmsisDapError::UnexpectedAnswer)).with_context(|| {
-                format!(
-                    "Connecting to target failed, received: {:x}",
-                    buffer[offset]
-                )
-            }),
+            _ => Err(SendError::ConnectResponseError(buffer[offset])),
         }
     }
 }