+++ probe-rs/src/probe/cmsisdap/commands/general/disconnect.rs
@@ -1,4 +1,4 @@
-use super::super::{Category, Request, Response, Result, Status};
+use super::super::{Category, Request, Response, SendError, Status};
 
 #[derive(Clone, Copy, Debug)]
 pub struct DisconnectRequest;
@@ -6,7 +6,7 @@ pub struct DisconnectRequest;
 impl Request for DisconnectRequest {
     const CATEGORY: Category = Category(0x03);
 
-    fn to_bytes(&self, _buffer: &mut [u8], _offset: usize) -> Result<usize> {
+    fn to_bytes(&self, _buffer: &mut [u8], _offset: usize) -> Result<usize, SendError> {
         Ok(0)
     }
 }
@@ -15,7 +15,7 @@ impl Request for DisconnectRequest {
 pub struct DisconnectResponse(pub(crate) Status);
 
 impl Response for DisconnectResponse {
-    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self> {
+    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self, SendError> {
         Ok(DisconnectResponse(Status::from_byte(buffer[offset])?))
     }
 }