+++ probe-rs/src/probe/cmsisdap/commands/swd/configure.rs
@@ -1,4 +1,4 @@
-use super::super::{Category, Request, Response, Result, Status};
+use super::super::{Category, Request, Response, SendError, Status};
 
 #[derive(Debug)]
 pub struct ConfigureRequest;
@@ -6,7 +6,7 @@ pub struct ConfigureRequest;
 impl Request for ConfigureRequest {
     const CATEGORY: Category = Category(0x13);
 
-    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize> {
+    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize, SendError> {
         // TODO: Allow configuration
         buffer[offset] = 0;
         Ok(1)
@@ -17,7 +17,7 @@ impl Request for ConfigureRequest {
 pub struct ConfigureResponse(pub(crate) Status);
 
 impl Response for ConfigureResponse {
-    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self> {
+    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self, SendError> {
         Ok(ConfigureResponse(Status::from_byte(buffer[offset])?))
     }
 }