+++ probe-rs/src/probe/cmsisdap/commands/swj/clock.rs
@@ -1,5 +1,4 @@
-use super::super::{Category, Request, Response, Result, Status};
-use anyhow::anyhow;
+use super::super::{Category, Request, Response, SendError, Status};
 
 #[derive(Debug)]
 pub struct SWJClockRequest(pub(crate) u32);
@@ -7,12 +6,12 @@ pub struct SWJClockRequest(pub(crate) u32);
 impl Request for SWJClockRequest {
     const CATEGORY: Category = Category(0x11);
 
-    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize> {
+    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize, SendError> {
         use scroll::{Pwrite, LE};
 
         buffer
             .pwrite_with(self.0, offset, LE)
-            .map_err(|_| anyhow!("This is a bug. Please report it."))?;
+            .map_err(|_| SendError::Bug)?;
         Ok(4)
     }
 }
@@ -21,7 +20,7 @@ impl Request for SWJClockRequest {
 pub(crate) struct SWJClockResponse(pub(crate) Status);
 
 impl Response for SWJClockResponse {
-    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self> {
+    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self, SendError> {
         Ok(SWJClockResponse(Status::from_byte(buffer[offset])?))
     }
 }