+++ probe-rs/src/probe/cmsisdap/commands/swj/pins.rs
@@ -1,4 +1,4 @@
-use super::super::{Category, Request, Response, Result};
+use super::super::{Category, Request, Response, SendError};
 
 pub struct SWJPinsRequest {
     /// A mask of the values the different pins selected in the selection mask will be set to.
@@ -114,7 +114,7 @@ bitfield::bitfield! {
 impl Request for SWJPinsRequest {
     const CATEGORY: Category = Category(0x10);
 
-    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize> {
+    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize, SendError> {
         use scroll::{Pwrite, LE};
 
         buffer
@@ -131,7 +131,7 @@ impl Request for SWJPinsRequest {
 }
 
 impl Response for Pins {
-    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self> {
+    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self, SendError> {
         Ok(Pins(buffer[offset]))
     }
 }