+++ probe-rs/src/probe/cmsisdap/commands/transfer/configure.rs
@@ -1,5 +1,4 @@
-use super::super::{Category, Request, Response, Result, Status};
-use anyhow::anyhow;
+use super::super::{Category, Request, Response, SendError, Status};
 
 /// The DAP_TransferConfigure Command sets parameters for DAP_Transfer and DAP_TransferBlock.
 #[derive(Debug)]
@@ -15,16 +14,16 @@ pub struct ConfigureRequest {
 impl Request for ConfigureRequest {
     const CATEGORY: Category = Category(0x04);
 
-    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize> {
+    fn to_bytes(&self, buffer: &mut [u8], offset: usize) -> Result<usize, SendError> {
         use scroll::{Pwrite, LE};
 
         buffer[offset] = self.idle_cycles;
         buffer
             .pwrite_with(self.wait_retry, offset + 1, LE)
-            .map_err(|_| anyhow!("This is a bug. Please report it."))?;
+            .map_err(|_| SendError::Bug)?;
         buffer
             .pwrite_with(self.match_retry, offset + 3, LE)
-            .map_err(|_| anyhow!("This is a bug. Please report it."))?;
+            .map_err(|_| SendError::Bug)?;
         Ok(5)
     }
 }
@@ -33,7 +32,7 @@ impl Request for ConfigureRequest {
 pub struct ConfigureResponse(pub(crate) Status);
 
 impl Response for ConfigureResponse {
-    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self> {
+    fn from_bytes(buffer: &[u8], offset: usize) -> Result<Self, SendError> {
         Ok(ConfigureResponse(Status::from_byte(buffer[offset])?))
     }
 }